clockdiff: Fix print time_t problem on 32 bits musl
Similarly to 0fd2db7 this fixes warning on 32 bit musl 1.2.0, which
changed time_t as 64-bit across all archs [1]. This fixes warning:

../clockdiff.c:626:35: warning: format '%ld' expects argument of type
'long int', but argument 2 has type 'time_t' {aka 'long long int'}
[-Wformat=]
  626 |  printf("%ld %d %d\n", now, ctl.measure_delta, ctl.measure_delta1);
      |          ~~^           ~~~
      |            |           |
      |            long int    time_t {aka long long int}
      |          %lld

[1] https://musl.libc.org/time64.html

+++ clockdiff.c
@@ -623,7 +623,7 @@ int main(int argc, char **argv)
 				ctl.hisname, ctl.rtt, ctl.rtt_sigma, ctl.min_rtt,
 				ctl.measure_delta, ctl.measure_delta1, s);
 		} else {
-			printf("%ld %d %d\n", now, ctl.measure_delta, ctl.measure_delta1);
+			printf("%lld %d %d\n", (long long)now, ctl.measure_delta, ctl.measure_delta1);
 		}
 	}
 	exit(0);
